+++ src/Bugfree.Spo.Analytics.Cli/Database.fs
@@ -2,7 +2,6 @@
 
 open System
 open System.Data.SqlClient
-open System.Text.RegularExpressions
 open FSharp.Data
 open Bugfree.Spo.Analytics.Cli.Utils