+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Bugfree.Spo.Analytics 1.3.1 (2017-08-07)
+
+* Added logging around mailbox processor code to log every processing related exception ([#8](https://github.com/ronnieholm/Bugfree.Spo.Analytics/issues/8))
+
 ## Bugfree.Spo.Analytics 1.3.0 (2017-04-30)
 
 * Switched to Argu for command-line parsing ([#7](https://github.com/ronnieholm/Bugfree.Spo.Analytics/issues/7))